chore: Move around all classes to correct packages, from reviews
TooltipModel -> TooltipHandler
Tooltip.type package
AbstractItemInfoDeserializer -> WynnItem
