+++ common/src/main/java/com/wynntils/handlers/tooltip/gear/GearTooltipFooter.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.tooltip.gear;
+package com.wynntils.handlers.tooltip.gear;
 
 import com.wynntils.core.text.StyledText;
 import com.wynntils.models.elements.type.Powder;
