+++ common/src/main/java/com/wynntils/handlers/tooltip/type/IdentifiableItemInfo.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.tooltip;
+package com.wynntils.handlers.tooltip.type;
 
 import com.wynntils.models.character.type.ClassType;
 import com.wynntils.models.stats.type.StatActualValue;
