+++ common/src/main/java/com/wynntils/models/gear/GearInfoRegistry.java
@@ -14,7 +14,6 @@ import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Managers;
 import com.wynntils.core.net.Download;
 import com.wynntils.core.net.UrlId;
-import com.wynntils.models.gear.type.AbstractItemInfoDeserializer;
 import com.wynntils.models.gear.type.GearInfo;
 import com.wynntils.models.gear.type.GearMetaInfo;
 import com.wynntils.models.gear.type.GearRequirements;
@@ -23,6 +22,7 @@ import com.wynntils.models.gear.type.GearType;
 import com.wynntils.models.stats.type.FixedStats;
 import com.wynntils.models.stats.type.StatPossibleValues;
 import com.wynntils.models.stats.type.StatType;
+import com.wynntils.models.wynnitem.AbstractItemInfoDeserializer;
 import com.wynntils.utils.JsonUtils;
 import com.wynntils.utils.type.Pair;
 import java.lang.reflect.Type;
