+++ common/src/main/java/com/wynntils/models/ingredients/IngredientInfoRegistry.java
@@ -18,11 +18,11 @@ import com.wynntils.core.net.Download;
 import com.wynntils.core.net.UrlId;
 import com.wynntils.core.net.event.NetResultProcessedEvent;
 import com.wynntils.models.elements.type.Skill;
-import com.wynntils.models.gear.type.AbstractItemInfoDeserializer;
 import com.wynntils.models.ingredients.type.IngredientInfo;
 import com.wynntils.models.ingredients.type.IngredientPosition;
 import com.wynntils.models.profession.type.ProfessionType;
 import com.wynntils.models.stats.type.StatType;
+import com.wynntils.models.wynnitem.AbstractItemInfoDeserializer;
 import com.wynntils.models.wynnitem.type.ItemMaterial;
 import com.wynntils.utils.JsonUtils;
 import com.wynntils.utils.type.Pair;
