+++ common/src/main/java/com/wynntils/models/rewards/TomeInfoRegistry.java
@@ -15,7 +15,6 @@ import com.wynntils.core.components.Managers;
 import com.wynntils.core.net.Download;
 import com.wynntils.core.net.UrlId;
 import com.wynntils.models.elements.type.Skill;
-import com.wynntils.models.gear.type.AbstractItemInfoDeserializer;
 import com.wynntils.models.gear.type.GearDropRestrictions;
 import com.wynntils.models.gear.type.GearMetaInfo;
 import com.wynntils.models.gear.type.GearRestrictions;
@@ -25,6 +24,7 @@ import com.wynntils.models.rewards.type.TomeRequirements;
 import com.wynntils.models.rewards.type.TomeVariant;
 import com.wynntils.models.stats.type.StatPossibleValues;
 import com.wynntils.models.stats.type.StatType;
+import com.wynntils.models.wynnitem.AbstractItemInfoDeserializer;
 import com.wynntils.models.wynnitem.type.ItemMaterial;
 import com.wynntils.models.wynnitem.type.ItemObtainInfo;
 import com.wynntils.utils.JsonUtils;
