+++ common/src/main/java/com/wynntils/screens/guides/gear/GuideGearItemStack.java
@@ -4,12 +4,13 @@
  */
 package com.wynntils.screens.guides.gear;
 
+import com.wynntils.core.components.Handlers;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.components.Services;
+import com.wynntils.handlers.tooltip.TooltipBuilder;
 import com.wynntils.models.gear.type.GearInfo;
 import com.wynntils.models.items.WynnItemData;
 import com.wynntils.models.items.items.game.GearItem;
-import com.wynntils.models.tooltip.TooltipBuilder;
 import com.wynntils.screens.guides.GuideItemStack;
 import java.util.ArrayList;
 import java.util.List;
@@ -63,7 +64,7 @@ public final class GuideGearItemStack extends GuideItemStack {
     }
 
     public void buildTooltip() {
-        TooltipBuilder tooltipBuilder = Models.Tooltip.buildNew(gearInfo, null, true);
+        TooltipBuilder tooltipBuilder = Handlers.Tooltip.buildNew(gearInfo, null, true);
         this.generatedTooltip = tooltipBuilder.getTooltipLines(Models.Character.getClassType());
 
         // Force ItemStatInfoFeature to recreate its cache
